+++ mediation/AppLovin/source/plugin/Assets/GoogleMobileAds/Platforms/Android/Mediation/AppLovin/AppLovinClient.cs
@@ -42,6 +42,15 @@ namespace GoogleMobileAds.Android.Mediation.AppLovin
             AndroidJavaClass appLovin = new AndroidJavaClass("com.applovin.sdk.AppLovinSdk");
             appLovin.CallStatic("initializeSdk", currentActivity);
         }
+
+        public void SetHasUserConsent(bool hasUserConsent)
+        {
+            MonoBehaviour.print("Updating AppLovin user consent flag");
+            AndroidJavaClass unityPlayer = new AndroidJavaClass("com.unity3d.player.UnityPlayer");
+            AndroidJavaObject currentActivity = unityPlayer.GetStatic<AndroidJavaObject>("currentActivity");
+            AndroidJavaClass appLovinPrivacySettings = new AndroidJavaClass("com.applovin.sdk.AppLovinPrivacySettings");
+            appLovinPrivacySettings.CallStatic("setHasUserConsent", hasUserConsent, currentActivity);
+        }
     }
 }