+++ mediation/AppLovin/CHANGELOG.md
@@ -1,5 +1,14 @@
 # AppLovin Adapter plugin for Google Mobile Ads SDK for Unity 3D Changelog
 
+## 3.0.0
+- Supports AppLovin Android SDK version 8.0.0.
+- Supports AppLovin iOS SDK version 5.0.1.
+- Added `AppLovin.SetHasUserConsent()` method to forward user consent flag to the AppLovin SDK.
+
+## 2.0.0
+- Supports AppLovin Android SDK version 7.8.6.
+- Supports AppLovin iOS SDK version 5.0.1.
+
 ## 1.2.1
 - Supports AppLovin Android SDK version 7.8.6.
 - Supports AppLovin iOS SDK version 4.8.3.